Remove superfluous and non-compliant YAML syntax from workflow
Previously, the "flow style" YAML syntax was used for this mapping. That was inconsistent with the rest of the workflow
code, which uses the "block style" that is the standard for all YAML code in Arduino projects. This type of random
mixing of the two styles makes otherwise fairly intuitive YAML code difficult to understand and maintain for anyone
without deep familiarity with the obscure details of this language.

Worse, the syntax had a form that was not specification-compliant:

https://yaml.org/spec/1.2.2/#823-block-nodes

> Flow nodes must be indented by at least one more space than the parent block collection.

Although some lenient parsers do happen to support the previous non-compliant form, such support is not guaranteed, and
non-compliance will cause a failure when used with any strict parser.