Remove unused `platforms` action input from "Compile Examples" workflow
The "arduino/compile-sketches" action provides a `platforms` input by which the platform dependencies to be installed
may be specified.

This input is useful for use cases such as:

* a 3rd party platform dependency which requires a supplemental package index
* a platform dependency from an alternative source (e.g., development version of a platform from a Git repo)
* transitory platform dependencies
* version pinning

However, the input is not required in the case where the platform is available from the primary package index, since the
action will derive the platform ID from the FQBN.

The workflow was based on a workflow from another repository which did explicitly define platform dependencies, and so
did require the use of the `platform` input. However, the platform dependencies data was removed from this workflow, so
the `platform` input serves no purpose and its presence only makes the workflow more difficult to understand and
maintain.