Remove unused `platforms` action input from "Compile Examples" workflow (#9)
* Remove superfluous and non-compliant YAML syntax from workflow

Previously, the "flow style" YAML syntax was used for this mapping. That was inconsistent with the rest of the workflow
code, which uses the "block style" that is the standard for all YAML code in Arduino projects. This type of random
mixing of the two styles makes otherwise fairly intuitive YAML code difficult to understand and maintain for anyone
without deep familiarity with the obscure details of this language.

Worse, the syntax had a form that was not specification-compliant:

https://yaml.org/spec/1.2.2/#823-block-nodes

> Flow nodes must be indented by at least one more space than the parent block collection.

Although some lenient parsers do happen to support the previous non-compliant form, such support is not guaranteed, and
non-compliance will cause a failure when used with any strict parser.

* Remove unused `platforms` action input from "Compile Examples" workflow

The "arduino/compile-sketches" action provides a `platforms` input by which the platform dependencies to be installed
may be specified.

This input is useful for use cases such as:

* a 3rd party platform dependency which requires a supplemental package index
* a platform dependency from an alternative source (e.g., development version of a platform from a Git repo)
* transitory platform dependencies
* version pinning

However, the input is not required in the case where the platform is available from the primary package index, since the
action will derive the platform ID from the FQBN.

The workflow was based on a workflow from another repository which did explicitly define platform dependencies, and so
did require the use of the `platform` input. However, the platform dependencies data was removed from this workflow, so
the `platform` input serves no purpose and its presence only makes the workflow more difficult to understand and
maintain.